- Added ROS_LOGGER prints to various classes - Improved GPS publisher - Changed ulisse_driver launch file

+++ ulisse_driver/include/ulisse_driver/CSerialHelper.h
@@ -57,7 +57,7 @@ public:
 
 	void ChangeBaudRate(uint32_t baudRate);
 
-	int Write(const char *buffer, int size);
+    ssize_t Write(const char *buffer, size_t size);
 	int SetRTS(int level);
 	int SetDTR(int level);
 	int SetCD(int level);
