+++ ulisse_driver/src/ulisse_driver/CSerialHelper.cc
@@ -196,18 +196,17 @@ void CSerialHelper::ChangeBaudRate(uint32_t baudRate) {
 	pthread_mutex_unlock(&critSecSem_);
 }
 
-int CSerialHelper::Write(const char *buffer, int size) {
-	int ret = SERIAL_ERROR;
-	int sent = 0;
+ssize_t CSerialHelper::Write(const char *buffer, size_t size) {
+    ssize_t ret = SERIAL_ERROR;
+    int sent = 0;
 
-	while (sent < size) {
+    while (sent < (int)size) {
 		ret = write(this->fd_, &buffer[sent], size - sent);
 		if (ret > 0) {
 			fsync(this->fd_);
 			sent += ret;
 		} else if (ret == -1) {
             printf("CSerialHelper::Write, Error on serial writing\n");
-
 			return SERIAL_ERROR;
 		}
 	}
